Notes About The Regex Edit In `build-lib.sh`
We have tried to use positive / negative "lookahead"

With the following examples:
```
AnonymousObject.a
AnonymousObject2.a
AnonymousObject3.a
AnonymousObject23.a
AnonymousObject32.a
AnonymousObject32.b
AnonymousObject33.a
AnonymousObject32abc.a
AnonymousObject32abc.b
```
https://regexr.com/ is a good website for testing.

The general answer for doing so was here:
https://stackoverflow.com/a/8240998/14427765

And in our case, the regex for receiving the 64bit files
would be the following:
```
(?=\b\w+)(?!.*32.*\.a).*\.a
```

But it seems that the `find` command does not support
regex "lookahead".

So we needed to find another option for solving the case.
(As we made in the previous commit.)

Signed-off-by: Tal Jacob <taljacob2@gmail.com>